+++ src/lightwin/core/list_of_elements/list_of_elements.py
@@ -294,15 +294,10 @@ class ListOfElements(list):
         Raises
         ------
         NotImplementedError
-            If ``which_phase`` is different from ``"phi_0_abs"`` or
-            ``"phi_0_rel"``.
+            If ``which_phase`` is ``"as_in_original_dat"``.
 
         """
-        if which_phase == "phi_s":
-            raise NotImplementedError(
-                "Output with SET_SYNC_PHASE not implemented yet."
-            )
-        if which_phase in ("as_in_settings", "as_in_original_dat"):
+        if which_phase in ("as_in_original_dat",):
             raise NotImplementedError
         self.files["dat_file"] = dat_file
         dat_filecontent = [